move rejecting fetch to middleware utils
