+++ src/middleware/sequence.js
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import rejectingFetch from './rejectingFetch';
+import rejectingFetch from './utils/rejectingFetch';
 import { getItem, setItem } from './localStorageCache';
 import { headersGet, headersPost } from './headers';
 import { dataApiPath } from './paths';
