+++ src/middleware/utils/uploadFiles.js
@@ -13,7 +13,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-import rejectingFetch from '../rejectingFetch';
+import rejectingFetch from './rejectingFetch';
 import invariant from 'invariant';
 import { isFile } from './fileReader';
